api updates

+++ NBP_Project_2023/Shared/PostOffice.cs
@@ -11,10 +11,10 @@ namespace NBP_Project_2023.Shared
         public int Id { get; set; }
         public string City { get; set; } = string.Empty;
         public int PostalCode { get; set; }
-        public double PostX { get; set; }
-        public double PostY { get; set; }
+        public float PostX { get; set; }
+        public float PostY { get; set; }
         public bool IsMainPostOffice { get; set; } = false;
-        public List<int> Workers { get; set; } = new List<int>(); // Courier Id
+        public List<int> Workers { get; set; } = new List<int>(); // CourierIds
         public List<string> Packages { get; set; } = new List<string>(); //PackageID
     }
 }
